api/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
requests>=2.31.0

